Move config migration from post-install to src
