+++ post-install.js
@@ -5,129 +5,14 @@ import fs from 'node:fs';
 import path from 'node:path';
 import process from 'node:process';
 import yaml from 'yaml';
-import _ from 'lodash';
 import chalk from 'chalk';
 import { createRequire } from 'node:module';
+import { addMissingConfigValues } from './src/config-init.js';
 
 /**
  * Colorizes console output.
  */
 const color = chalk;
-
-const keyMigrationMap = [
-    {
-        oldKey: 'disableThumbnails',
-        newKey: 'thumbnails.enabled',
-        migrate: (value) => !value,
-    },
-    {
-        oldKey: 'thumbnailsQuality',
-        newKey: 'thumbnails.quality',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'avatarThumbnailsPng',
-        newKey: 'thumbnails.format',
-        migrate: (value) => (value ? 'png' : 'jpg'),
-    },
-    {
-        oldKey: 'disableChatBackup',
-        newKey: 'backups.chat.enabled',
-        migrate: (value) => !value,
-    },
-    {
-        oldKey: 'numberOfBackups',
-        newKey: 'backups.common.numberOfBackups',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'maxTotalChatBackups',
-        newKey: 'backups.chat.maxTotalBackups',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'chatBackupThrottleInterval',
-        newKey: 'backups.chat.throttleInterval',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'enableExtensions',
-        newKey: 'extensions.enabled',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'enableExtensionsAutoUpdate',
-        newKey: 'extensions.autoUpdate',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'extras.disableAutoDownload',
-        newKey: 'extensions.models.autoDownload',
-        migrate: (value) => !value,
-    },
-    {
-        oldKey: 'extras.classificationModel',
-        newKey: 'extensions.models.classification',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'extras.captioningModel',
-        newKey: 'extensions.models.captioning',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'extras.embeddingModel',
-        newKey: 'extensions.models.embedding',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'extras.speechToTextModel',
-        newKey: 'extensions.models.speechToText',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'extras.textToSpeechModel',
-        newKey: 'extensions.models.textToSpeech',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'minLogLevel',
-        newKey: 'logging.minLogLevel',
-        migrate: (value) => value,
-    },
-    {
-        oldKey: 'cardsCacheCapacity',
-        newKey: 'performance.memoryCacheCapacity',
-        migrate: (value) => `${value}mb`,
-    },
-    {
-        oldKey: 'cookieSecret',
-        newKey: 'cookieSecret',
-        migrate: () => void 0,
-        remove: true,
-    },
-];
-
-/**
- * Gets all keys from an object recursively.
- * @param {object} obj Object to get all keys from
- * @param {string} prefix Prefix to prepend to all keys
- * @returns {string[]} Array of all keys in the object
- */
-function getAllKeys(obj, prefix = '') {
-    if (typeof obj !== 'object' || Array.isArray(obj) || obj === null) {
-        return [];
-    }
-
-    return _.flatMap(Object.keys(obj), key => {
-        const newPrefix = prefix ? `${prefix}.${key}` : key;
-        if (typeof obj[key] === 'object' && !Array.isArray(obj[key])) {
-            return getAllKeys(obj[key], newPrefix);
-        } else {
-            return [newPrefix];
-        }
-    });
-}
 
 /**
  * Converts the old config.conf file to the new config.yaml format.
@@ -152,71 +37,6 @@ function convertConfig() {
             console.error(color.red('FATAL: Config conversion failed. Please check your config.conf file and try again.'), error);
             return;
         }
-    }
-}
-
-/**
- * Compares the current config.yaml with the default config.yaml and adds any missing values.
- */
-function addMissingConfigValues() {
-    try {
-        const defaultConfig = yaml.parse(fs.readFileSync(path.join(process.cwd(), './default/config.yaml'), 'utf8'));
-        let config = yaml.parse(fs.readFileSync(path.join(process.cwd(), './config.yaml'), 'utf8'));
-
-        // Migrate old keys to new keys
-        const migratedKeys = [];
-        for (const { oldKey, newKey, migrate, remove } of keyMigrationMap) {
-            if (_.has(config, oldKey)) {
-                if (remove) {
-                    _.unset(config, oldKey);
-                    migratedKeys.push({
-                        oldKey,
-                        newValue: void 0,
-                    });
-                    continue;
-                }
-
-                const oldValue = _.get(config, oldKey);
-                const newValue = migrate(oldValue);
-                _.set(config, newKey, newValue);
-                _.unset(config, oldKey);
-
-                migratedKeys.push({
-                    oldKey,
-                    newKey,
-                    oldValue,
-                    newValue,
-                });
-            }
-        }
-
-        // Get all keys from the original config
-        const originalKeys = getAllKeys(config);
-
-        // Use lodash's defaultsDeep function to recursively apply default properties
-        config = _.defaultsDeep(config, defaultConfig);
-
-        // Get all keys from the updated config
-        const updatedKeys = getAllKeys(config);
-
-        // Find the keys that were added
-        const addedKeys = _.difference(updatedKeys, originalKeys);
-
-        if (addedKeys.length === 0 && migratedKeys.length === 0) {
-            return;
-        }
-
-        if (addedKeys.length > 0) {
-            console.log('Adding missing config values to config.yaml:', addedKeys);
-        }
-
-        if (migratedKeys.length > 0) {
-            console.log('Migrating config values in config.yaml:', migratedKeys);
-        }
-
-        fs.writeFileSync('./config.yaml', yaml.stringify(config));
-    } catch (error) {
-        console.error(color.red('FATAL: Could not add missing config values to config.yaml'), error);
     }
 }
 
@@ -288,7 +108,7 @@ try {
     // 1. Create default config files
     createDefaultFiles();
     // 2. Add missing config values
-    addMissingConfigValues();
+    addMissingConfigValues(path.join(process.cwd(), './config.yaml'));
 } catch (error) {
     console.error(error);
 }
