+++ src/transformers.js
@@ -12,7 +12,6 @@ configureTransformers();
 function configureTransformers() {
     // Limit the number of threads to 1 to avoid issues on Android
     env.backends.onnx.wasm.numThreads = 1;
-    console.log(env.backends.onnx.wasm.wasmPaths);
     // Use WASM from a local folder to avoid CDN connections
     env.backends.onnx.wasm.wasmPaths = path.join(serverDirectory, 'node_modules', 'sillytavern-transformers', 'dist') + path.sep;
 }
